Deprecate string inflection methods, require 'sequel/extensions/inflector' if you use them
This commit adds a Sequel::Inflections module that extends Model and
is included in AssociationReflection.  It basically does the work
of the old string inflection code, but without modifying String.

This commit leaves us in a less than perfect state, which we'll be in
until the deprecated inflector is removed.  If the new inflector
extension is required, there will be duplicate entries in the various
internal variables.  That shouldn't change any output, though it may
cause slowdown.

In order to preserve backwards compatibility as much as possible,
both the deprecated inflector and the new extension update Sequel's
internal copy.  After the deprecated inflector is removed, I'll
be changing the new extension not to update Sequel's copy, and have
Sequel call the string inflector methods if they are available. I
didn't do this now because it would result in calling the
deprecated inflector methods even when doing something that isn't
deprecated.

This commit also renames hooks.rb to deprecated_hooks.rb, for
consistency with the other deprecated files.