Implement base model functionality and associations as model plugins
This commit makes plugins required, as it makes Sequel::Model and
Sequel::Model::Associations plugins.  This shows how powerful the
model plugin feature is, since it implements even the most basic
model functionality as a plugin.

This commit moves class and instance methods related to associations
to the new associations plugin.  The old EagerLoading module is
now Associations::DatasetMethods.

If plugins include DatasetMethods, it is no longer an error to apply
them on a class that doesn't have a dataset.  Sequel will extend
the dataset with all plugin modules inside set_dataset.

In order to keep each plugin self contained to a single file,
record.rb and dataset_methods.rb were merged into base.rb, and
eager_loading.rb was merged into associations.rb.

This commit makes plugin the only singleton method of Model.  This
is to make it easier to have the basic functionality implemented as
a plugin.