Deprecate Database#transaction accepting a server symbol argument, make PostgreSQL savepoint support optional
This commit is designed to make it easier for methods that call
Database#transaction to give it options.  The change was mainly made
to support making savepoints optional in the PostgreSQL adapter.

The PostgreSQL adapter used to always use savepoints.  However, that
made Sequel's behavior for the following type of code different
depending on whether you were using PostgreSQL:

  class Author < Sequel::Model
    one_to_many :books
    def add_book(b)
      raise Sequel::Error::Rollback if /bad/.match(b.name)
      super
    end

    def add_all_books(books)
      db.transaction{books.each{|book| add_book(book)}}
    end

    def steal_books_from(author)
      db.transaction do
        add_all_books(author.books)
        author.remove_all_books
      end
    end
  end
  Author[1].steal_books_from(Author[2])

On SQLite, if author 2 has a book with a bad name, the
steal_books_from call will not cause any database changes.  However,
if you run it on PostgreSQL, if will remove all books from author 2
without adding them to author 1.  That's not generally what you want
to happen.

I think savepoints should be an explicit choice on the user's part,
so on PostgreSQL (the only database that supported savepoints), you
now have to use the :savepoint option to Database#transaction to
use a savepoint, otherwise, it will reuse the existing
transaction/savepoint.

I have to give ActiveRecord credit for doing the right thing in this
situation first, by making the savepoint support optional when adding
it.  I didn't consider the negative ramifications when I merged the
PostgreSQL savepoint patch.