Fix ~{:bool_col=>true} and related inversions of boolean values
Before using true or false as a hash value was not treated specially,
it was treated as an equals relationship.  That doesn't mesh well
with SQL's 3 valued boolean logic.  SQL provides a way to deal with
boolean values, IS [NOT] (TRUE|FALSE).  Unfortunately, Sequel didn't
use that construction (until now).

It generally makes no difference in the IS TRUE or IS FALSE case, but
it makes a difference in the IS NOT TRUE and IS NOT FALSE case.
IS NOT (TRUE|FALSE) should return NULL values, but Sequel was using
!= '(t|f)', which doesn't return NULL values.

The change is fairly straight forward, using the IS and IS NOT
operator for TrueClass and FalseClass in addition to NilClass.
The output needs to be special-cased in complex_expression_sql
though, as you need to use TRUE and FALSE instead of 't' and 'f'.

IS [NOT] (TRUE|FALSE) is in SQL92, but SQLite doesn't support it,
work around that by using != '(t|f)' OR IS NULL.

I added an integration test to check for this.