Get rid of Sequel::Schema::SQL, move methods into Database
The methods are database methods, and they are overridden in the
database subclasses in the adapters, so there isn't a reason to
have them in a separate module.

This commit moves some files around to get rid of the
sequel_core/schema directory.  As the schema generators are
really only used by create_table and alter_table, they make sense
to store in the sequel_core/database directory along with
the schema methods.  Rename database/schema.rb to
database/schema_methods.rb to make it more clear what is defined
inside.

Getting rid of the module required some small changes in the
adapters and documentation.

While here, add RDoc for some private methods in the generators.