Merge sequel_core directory into sequel directory
This is the first half of the merge, the merge of sequel_model into
sequel will come next.  I'm going this because it makes things
cleaner.  I've wanted to do this for a while, but it wasn't worth
the minor breakage.  Now that Sequel 3.0 is coming soon, it seemed
like a good time.

To make life easier and to allow for more flexibility, I'm adding
a Sequel.require method that simplifies requiring multiple files.
It also only loads files in the same directory or a subdirectory
of the directory containing lib/sequel/core.rb.  Sequel used to do
this many moons ago, and I removed it to give users more flexibility
to override stuff.  I've come to the conclusion that removing it
caused more problems than it solved.  There have been a couple of
cases during the recent refactoring where I got obscure errors
because I deleted a file but still had it required somewhere, and it
loaded a file in a previous gem.  If people want to customize sequel,
they should fork the repository or unpack the gem to make their
modifications.

This requires simple changes to all adapters.  I hope I didn't screw
anything up, but it should be easy to fix if I did.

This also fixes a reference to Schema::SQL in the shared MySQL
adapter.