Make Dataset#where always affect the WHERE clause
Before, Dataset#where was an alias of Dataset#filter.  However,
Dataset#filter modifies the HAVING clause if the query already has
a HAVING clause.  Because it doesn't make sense for Dataset#where
to affect HAVING, make #where its own method.  Add a private
_filter method that #filter, #having, and #where call.

This allows you to modify the WHERE clause after a HAVING clause
has been added, which Sequel was previously unable to do easily.

While here, fix the documentation, which said that filter
modified the HAVING clause if the query had been grouped, when it
only modified the HAVING clause if there was already a HAVING
clause.  I thought about making the code match the documentation,
but as there is an explicit spec that indicates that filter should
modify where even if the dataset is grouped, I decided it would be
better to change the documentation to match reality.