Don't have deprecated methods take precedence over methods added before or after
This changes the deprecated methods so that most methods are included
in modules instead of being added directly to classes, so they don't
override methods added later via module inclusion.  They also check
for existing method definitions, so they won't be added if something
else already defines the method.  This is much nicer, especially for
the String inflection methods and other methods that are also used
by other applications.