Change Database @@adapters to Sequel::ADAPTER_MAP
Using a constant instead of a class variables for a shared hash is
a much better idea.  Plus, this makes it easier for running code
to see which adapters have been loaded.

This does break backwards compatibility slightly, but there wasn't
a method that accessed the class variable, so it should have been
considered private anyway.