Deprecate Database #multi_threaded? and #logger, Update RDoc in database.rb
multi_threaded? was just the opposite of single_threaded, and it's
probably not used enough to justify it's own method.  logger has just
existed for backwards compatibility since I added support for
multiple loggers.  As it isn't accurate (only gives you the first
logger if you have multiple loggers), I'm deprecating it.  You can
still use the logger= method to set a single logger for the database.

I've also removed the ability of Database#[] to take a block. As you
can't use a block directly with the array access operator, and []
just calls either fetch or from, it's not worth it (I'm the one
who added support for it originally, I think).