Denamespace Sequel::Error decendants
This moves the Sequel's exception classes that were stored in the
Sequel::Error namespace to the Sequel namespace.  This makes code
that uses them less verbose.  For backwards compatibility, the old
names were kept.  I tested this by changing the lib code and
running the old specs, but this commit also includes spec updates.
The old names are not officially deprecated, but their use is
discouraged.

Sequel::Error::InvalidTransform and Sequel::Error::NoExistingFilter
do not have an equivalent in the Sequel namespace.  Both are
deprecated and will be removed in Sequel 3.0.  Dataset transforms
will be deprecated soon and removed in 3.0, so there isn't a need
for the related exception class.  Sequel::Error::NoExistingFilter
was only used in 2 places, which I converted to use
Sequel::InvalidOperation, which is used for similar reasons
elsewhere.  It's not as explicit, but I don't think we need
a separate exception class for the rare case.  I'm against
exception class proliferation in general, and I think there should
only be a separate exception class if you could conceivably want
to catch just that specific type of exception.

Note that Sequel::Error::PoolTimeoutError changed to
Sequel::PoolTimeout, I renamed it because I don't think the extra
Error at the end adds any value.

I also brought back Sequel::Error::InvalidStatement in deprecated
form, it will be removed in 3.0.

This commit also includes RDoc updates for exceptions.rb.