Deprecate Migration and Migrator, require 'sequel/extensions/migration' if you want them
Migrations and the Migrator are not generally needed in normal
operation.  The only part of Sequel that depends on them is the
sequel binary when called with the -m switch.  So it is easy to
move them into an extension.  Since migrations are only run rarely,
it doesn't make sense for Sequel to load the code for them by
default.