Add Sequel::MySQL.convert_invalid_date_time accessor for dealing with dates like "0000-00-00" and times like "25:00:00"
This is based on epugh's patch, though I've changed it quite a bit.
This allows you to convert the invalid dates to nil by setting the
accessor to nil or :nil, or to return them as strings by setting the
accessor to :string.  The default is false, which tells Sequel to
raise an error if the database returns an invalid date.