Deprecate Dataset#transform and Model.serialize, and model serialization plugin
This is the last major feature that will be deprecated and removed in
3.0.  Nobody has responded that they need the dataset transform
feature, so that is being removed.  It's tied into a lot of different
places in Dataset, so I'm not going to be providing an extension that
adds it back.

Model.serialize relies on Dataset#transform, so it is being
deprecated as well.  The new serialization plugin takes a different
approach.  It overrides the column reader to deserialize the column
on request, and adds a before_save hook that serializes the
deserialized values before saving.  It should do the job and be
reasonably backward compatible.

This commit moves the requiring of bigdecimal/util, enumerator, and
yaml to deprecated.rb, since they will be removed in 3.0.

This commit makes Model#reload an actual method instead of an alias,
to make overriding it in a plugin easier (which the serialization
plugin does).

This commit also reuses the DATASET_METHOD_RE in
Model#def_column_accessor.