Split multi_insert into 2 methods with separate APIs, keep (deprecated) backwards compatibility
The API for multi_insert has always bothered me.  It's too complex,
and the method should have been split from the beginning.  import
was an alias to multi_insert, so I've split the method and used
import as the underlying method that takes an array columns and
an array or arrays of values or a dataset, and had multi_insert
call import.  Backwards compatibility was kept, but it's deprecated
and will be removed before 3.0.

While here, fix a very misleading RDoc example for multi_insert,
probably added by me about a year ago because I didn't understand
the method's API (using *args when you don't actually accept an
arbitrary number of arguments is a bad idea).