Don't use the model association plugin if SEQUEL_NO_ASSOCIATIONS constant or environment variable is defined
If you don't need associations for your application, Sequel won't
force them on you any more.  This should cut down on memory usage
when associations are not needed.

This commit removes the little remaining coupling between the base
model plugin and the caching, validation class methods, and
association plugins.

This commit moves the association_reflection.rb file into
associations.rb.  Associations could be just like any other plugin
if I wanted to change the namespace hierarchy, but I don't think it
is worth it.  Almost everyone using models is going to want
associations of some kind, this is for the rare soul that doesn't
want it.

This commit allows you to have associations for only certain
classes by doing:

  require 'sequel/model/associations'
  MyModel.plugin Sequel::Model::Associations

This commit probably causes some ruby -w issues because of the
removal of the code that sets @transform to nil.  Transforms are
going away soon, so I'm willing to make this change now, so I
don't forget to make it later.

In order to fix a ruby -w issue in validation class methods
plugin, the skip_superclass_validations? was added to the
plugin's class methods.