RDoc updates in remaining model files, small fixes to reduce memory use
The majority of the lines changed are due to using constants instead of
creating new string and regexp objects in every call (which need to be
garbage collected).

This also fixes Model::Errors#count to not create full_messages, so
it should be faster.

This should be the last of the RDoc related commits.