Fix rollback bug in Database#transaction in the PostgreSQL adapter
When I changed the transaction method to only use savepoints when
they were specifically requested, I broke the transaction method when
Sequel::Rollback was raised.  Scaringly, this wasn't caught be the
specs or the app tests I run on a regular basis.

I added an adapter spec for this, and also added an integration
test suite related to database transactions, mostly copied from the
PostgreSQL adapter spec.  Hopefully this will ensure that breakage is
caught earlier in the future (ideally before I commit).