Remove all methods and features deprecated in 2.12
This huge commit removes the deprecated code and deprecated specs
(over 300!).  It also makes some other minor changes.

Sequel.virtual_row_instance_eval is now true, but setting it to false
does nothing.  Using instance_eval for virtual row blocks that don't
accept arguments is now standard Sequel behavior.

Some error messages in convenience.rb were turned into constants to
reduce garbaged produced.  More work on this will probably be done in
the future.

Dataset#import was using each_slice, which was probably provided by
enumerator.  Change to using a loop and slicing manually.

The built in inflector now calls the string inflection methods if
the string responds to the method.  The inflector extension no longer
updates the built in inflector, since it no longer needs to.

Database#blank_object? calls blank? on the object if the object
responds to it.

The connection pool had a slight code refactoring to make things
easier to read.

The hook_class_methods plugin instance methods call super, so using
the plugin doesn't ignore previous instance level hooks.  Among other
things, this allows you to use the hook_class_methods plugin after
the caching plugin, which fixes #264.

Make the serialization plugin test require yaml, since it
uses yaml.  This only seems necessary on ruby 1.9.

Sequel::Deprecation is being moved to the extra directory.  I'm
moving it out of lib so it won't show up in the RDoc.  I'm not
removing it completely because I expect it may be used again
sometime in the future.

One thing that I realize that I should have officially deprecated
was the 4th argument to join_table being a table_alias instead of a
hash of options.  So Sequel will continue to support that.