Make Dataset#quote_identifier work with SQL::Identifiers
I want to support this now that virtual rows will be used more.
This allows you do to things like:

  dataset.select{sum(hours).as(hours)}

Before, you had to do:

  dataset.select{sum(hours).as(:hours)}

It only saves a character, but this change is made mainly because the
previous behavior was a little inconsistent.  You should be able to
use the virtual row method as both the function argument and alias.