Add Amalgalite adapter
Amalgalite is a ruby library that embeds SQLite in a ruby extension
without requiring a separate SQLite installation.

This adapter is full featured and passes all integration tests, and
all but one of the SQLite adapter tests (I could fix that failing
test, but I don't think it's worth it).

The only major issue with the amalgalite adapter is that it is
currently pretty slow (10 times slower in the integration tests,
40 times slower in the adapter specs).  This could have something
to do with the fact that the adapter reloads the entire schema
whenever there is a possibility that cached information could be
stale.  I didn't want to do that, but otherwise amalgalite can
give you weird NoMethodErrors.

As I don't use this in production, I'm not planning on working on
the performance issues.