Refactor create table and temporary table support
Handle overrides of constants in subclasses correctly.  Lexical scope
overrides class scope.  To reduce duplication, use
self.class.const_get.  Unfortunately, that doesn't work for shared
adapters that define the overrides in modules, so add methods in
shared adapters that need to override the constants.

Switch from use :temporary as the option name to just :temp.

Add create_table_sql for handling just the table creation statement.

Don't override create_table_sql_list in the MSSQL and MySQL adapters.
Switch MySQL to override create_table_sql, but call super to avoid
duplication.

Make create_table_sql_list indexes argument not be optional.  Now
that this is a private method, this shouldn't break anything, and the
only caller of this is create_table, which provides the argument.

Firebird appears to support temporary tables starting in version 2.1,
so make the Firebird adapter support it.