Refactor MySQL adapter tests, fix index type bug
MySQL's documentation states that since 5.0.60, you should be able
to use index types after the index columns, but my 5.0.75 MySQL
server gives me an error when I do that, so move it before the
ON table name.  This is apparently deprecated according to MySQL's
documentation (http://dev.mysql.com/doc/refman/5.0/en/create-index.html),
but deprecated and working is better than supported and not working.

Most of the MySQL adapter changes are dealing with actually testing
the DDL commands on the server, instead of just checking that the
syntax used matches.  This method of testing is more robust, and is
what allowed me to find the index type bug.  It's going to be needed
for some alter_table changes coming soon.