Fix altering a column name, type, default, or NULL/NOT NULL status on MySQL
Previously, altering a column's default or NULL/NOT NULL status did
not work on MySQL.  Altering a column's name required you specify
a type and lost any default/NOT NULL status.  Altering a column's
type also lost any default/NOT NULL status.

Fix this by parsing the schema for the table to figure out the
current type, default, and NULL/NOT NULL status.  Add specs to make
sure things work as expected.

Note that this isn't a perfect fix.  I'm not sure if the schema
parsing handles everything in the column definition, so it is
certainly possible that information may be lost.  As a workaround,
use a database that sucks less.