Change Dataset#exclude to invert given hash argument, not negate it
This change only has an affect when #exclude is called with a hash or
array of all two pairs.  Before, #exclude had the following behavior:

  dataset.exclude(:a=>1, :b=>2)
  # => SELECT * FROM test WHERE ((a != 1) AND (b != 2))

This negated the conditions instead of inverting them, which went
against the documentation.  Now, the above code would yield the
following SQL:

  SELECT * FROM test WHERE ((a != 1) OR (b != 2))

So exclude should return all rows that would not be matched by
filter with the same arguments (well, except for NULL issues).

I'm pretty sure I'm the one that modified exclude to do negation
instead of inversion, and I think that was a mistake.  This behavior
wasn't documented, though it was in the specs.  One reason I think
this feature is a bad idea is that it is inconsistent.  It would
only use negation if passed a single hash argument.  If you used a
block, it would invert.  Now with filter correctly handling multiple
arguments, it would be even more inconsistent.

This breaks backwards compatibility, you should definitely check your
code for exclude and if it is called with a hash with multiple
keys, you will probably need to change your code.