Refactor Schema Generators, some backwards incompatable changes
This changes the schema generators to store constraints separately
from columns.  This simplifies some of the logic and internal code.

The columns, constraints, and indexes instance methods were added,
the create_info instance method was removed.

In addition to storing constraints in a separate instance variable,
the type of constraint is now stored in :type instead of in
:constraint_type.

The generator is now passed directly into create_table_sql
and column_list_sql methods.  The create_table_sql_list method has
been removed.  The index_list_sql_list method was renamed to
index_sql_list.

One consequence of the changes is that now table constraints are
always after column definitions.

This commit has some nontrivial changes to the firebird adapter that
should be tested.