Refactor how generic types specified as ruby classes work
I don't really like modifying this code, as it is tied closely to
migrations, and I try not to break old migrations.  However, as the
generic type support was added recently and people were previously
told not use use modifiers if a generic types were used, this
shouldn't break things unless those warnings were ignored.

This commit changes the generic type support to use ruby methods
that are easily overridable in subclasses instead of looking up
the methods in a hash.  The new methods have the ability to look
at not just the class used as a type, but also any options specified
with the column.

By looking at the options, there is a lot more flexibility, since you
can have Strings with a :fixed option give a char type, something
you were not able to do with the generic class support previously.
There is also support for specifying a generic text type (String with
:text option), as well as a generic time type (Time with :only_time
option).

Because casting is tied to the type literal conversion, that support
changed as well.  Database#cast_type_literal was added and yields
an appropriate database type for casting purposes for the given type.

The Time class in MySQL now uses the datetime type instead of a
timestamp type.  I think I originally used timestamp because it had
a similar range to ruby's Time, but since there are ruby Time values
that the MySQL timestamp range doesn't cover, that was probably a
poor choice.

On PostgreSQL, specifying a String type with a size now works by
creating a varchar type.  You can specify :text=>false to get a
varchar(255) type.

The Firebird adapter had a special fix for database specific text
types, remove that in favor of using String, :text=>true to get
the special behavior.

This commit removes the Database#type_literal_base methods, in
favor of the type_literal_generic* methods

While here, change Schema::Generator to use methods instead of
instance variables where appropriate.