Add schema_dumper extension, for dumping schema of tables
This adds three functions to Schema::Generator: dump_columns,
dump_constraints, and dump_indexes.  The expectation is that if you
instance eval each inside of another instance, that instance will
be a copy of current instance.

This adds a Database#dump_table_schema method, which returns a string
representing a create_table method call that will recreate the table.
It isn't perfect, but it recognizes primary keys, defaults, and
NULL/NOT NULL determinations.  It also is able to parse database
types back into ruby types, and use those.  It converts types by
default, so if you wanted to copy the table schema from a PostgreSQL
database to an SQLite database, you could use it.  If you are copying
to the same database type, you can use the :same_db option to not
convert database types.

The conversion from database types to ruby types is lossy.  Most of
the time you will get the same or similar type, but if you have a
single precision float, it'll probably give you a double precision
float.  And if you have a smallint, you'll end up with a regular
integer.  These cases could be dealt with, but I'd like to keep the
code as simple as possible.

A Database#dump_schema_migration method was also added, that can be
used to dump the entire database as a single migration.

Currently this only supports dumping columns.  It will be expanded to
dump indexes as well soon.