Allow Dataset#from to remove existing FROM tables when called without an argument, instead of raising an error later
This change is backwards incompatible, but previously there was no
way to remove all FROM tables.  I chose to backwards compatibility
and make from consistent with select, in that still works if you
call it without an argument.  I thought about adding a method
similar to select_all, but couldn't think of a good name (from_none
was my original choice, but I don't really like that).