Add index parsing to PostgreSQL, MySQL, SQLite, and JDBC adapters, support dumping indexes
This commit adds a Database#indexes method to the above adapters,
which share the following API (from the RDoc):

  Return a hash containing index information. Hash keys are index
  name symbols. Values are subhashes with two keys, :columns and
  :unique.  The value of :columns is an array of symbols of column
  names.  The value of :unique is true or false depending on if the
  index is unique.

Here's an example of use:

  DB.indexes(:songs)
  => {:songs_name_index=>{:unique=>true, :columns=>[:name]},
      :songs_lyricid_index=>{:unique=>false, :columns=>[:lyricid]}}

This information is used by the schema_dumper extension to dump
indexes when dumping tables.  There will be an option later to skip
the dumping of indexes.

This also changes the schema_dumper extension to dump composite
primary key constraints, or really any primary key constraints that
isn't an autoincrementing primary key.

While here, I'm refactoring the tables and schema_parse_table methods
to make them more DRY.  To that end I'm adding a metadata_dataset
private method that uses the default identifier input and output
methods, as well as input_identifier_meth and output_identifier_meth,
both of which return method objects.  The point of all this is so
that things don't break when a user uses an input or output
identifier method different from the default.