Fix issues with the index parsing
The JDBC adapter was completely broken in the last commit, fix that.
The MySQL adapter's indexes method was broken, fix that as well.
The SQLite adapter's indexes method was broken on JDBC, handle that.

Add tests to make sure primary key indexes are not included, and
modify the SQLite, JDBC, and H2 adapters to make sure of that.