Allow schema_dumper extension to dump indexes separate from columns
You can now use the dump_indexes_migration method to dump just
indexes in a separate migration.  This is helpful when cloning
databases, as it is faster to transfer all data first and then
add indexes.

This includes some small refactoring in dump_table_schema.  It
also modified Generator#dump_indexes to accept :add_index and
:drop_index methods.  This also documents the :indexes option
to dump_schema_migration.