Make private select_*_sql methods only take a single argument
Now that select_sql doesn't take any options, we can streamline the
code so that instead of passing the @opts instance variable to
every method as an argument, we just refer to it inside the methods.

This is probably a little faster based on a few spec runs that I did,
but the main reason I'm doing this is that the code is cleaner.