Make Model#save use the same server it uses for saving as for retrieving the saved record
This change should only affect those who are using Sequel's
master/slave database functionality with Models.  Because
Model#save runs in a transaction by default, and because
Model#refresh does a select and Model#save does an insert for new
records, before, Model#save would insert the record on the :default
server and then refresh.  Since refresh does a read, it
operates on the :read_only server by default.  Because it was inside
a transaction, the changes from the master weren't available on the
slave yet.  Anyway, the fix is to use the :default server for
refreshes that are called inside of save.