Remove dead threads from connection pool if the pool is full and a connection is requested
Thanks to a report on IRC by Crash2burn, I was able to track down
this bug.  The bug is related to Sequel's connection pool's lack of
understanding of ruby's threading implementation. Sequel's connection
pool was apparently designed with the idea that using ensure will
ensure that the code is run.  However, that isn't true in all cases.
For example, if Thread#exit! is called, the thread terminates
immediately without running any ensure blocks.

One would assume that nobody would use Thread#exit! or a similar
method without a very good reason (looks like both jruby and ruby 1.9
don't support it at all), but apparently if ruby is used with many
threads, it will call that or something similar.  I was able to
recreate the issue using the following code:

  ds = DB[:table]
  1000.times{Thread.new{sleep 1; rand > 0.5 ? ds.insert : ds.first}}

Basically, what happens is that the connection pool's @allocated
hash is left with dead threads.  These should be released by the
ensure block, but apparently that isn't always run.

The fix is fairly simple.  If a connection request comes in and the
pool is full, check the @allocated hash for dead threads and release
any that are dead.

This approach is lazy in that it doesn't release dead thread
connections right away.  However, it will only keep the dead thread
around if you aren't fully utilizing the pool.  As soon as you start
to fully utilize the pool, it will clean up the garbage.  I expect
this has a modest performance impact, but I don't think it can be
done faster and still be correct.  Send me a patch if I'm wrong.

There's no change to the single threaded pool, as this situation
will only happen in multithreaded code.

While here, stop keeping track of the connection counts separately.
I'm not sure of the performance impact of that change, but it
simplifies the code and makes future bugs less likely.

This also moves the acquire call inside the begin/ensure block
in #hold, which should fix a problem if a thread died directly after
requiring a connection but before the begin block was entered.

This also changes #release so that it doesn't take a connection
object, and it now assumes the mutex is already in use.  I could
make a similar change to #remove, but there isn't any reason to
at this point.