Remove Sequel::Postgres::Adapter#reset, modify prepared statement variable initialization
The #reset method was used when the adapter detected a disconnection.
Now it raises a DatabaseDisconnectError, so the method is no longer
needed.

The prepared statement implementation is specific to each connection,
so there wasn't a thread safety issue with the old implementation.
The newer one should be a little simpler and negligibly faster.