Refactor the schema parsing data structure
This removes some complexity of the schema parsing implementation.
It used to be a hash with a default proc.  That was necessary in
order to be backwards compatible with the previous code that allowed
you to call schema without a table argument and get the schema for
all tables.  Since that was removed in 3.0, it's no longer needed.

While here, remove the optimization in the table_exists? method
where it checks the schema.  It could have been modified to work
with the new data structure, but I don't think it is likely to be
used in any preformance sensitive code.

I'm adding an integration test for table_exists? while I'm here too.