Handle ordered and limited datasets correctly when using UNION, INTERSECT, or EXCEPT
This changes the support for compound datasets in the following ways:

1) If the receiver dataset has an order or limit, it is now wrapped
   in a subselect.

2) If the given dataset (the argument) has an order or limit, it is
   now wrapped in a subselect.

3) The returned compound dataset is now wrapped in a subselect.

Example:

  DB[:items].limit(2).order(:a).
    union(DB[:old_items].limit(3).order(:b)).
    order(:c).limit(4)
  ### SQL
  # SELECT * FROM
  #  (SELECT * FROM
  #    (SELECT * FROM items ORDER BY a LIMIT 2) AS t1
  #     UNION
  #     SELECT * FROM
  #       (SELECT * FROM old_items ORDER BY b LIMIT 3) AS t1
  #   ) AS t1
  # ORDER BY c LIMIT 4

By always returning a dataset wrapped in a subselect, using
any other dataset call after a subselect call will always
operate on the full compound dataset itself instead of the first
dataset in the compound dataset (the previous receiver).

This should make ordering and limiting work intuitively with compound
datasets.

An integration test was added to check that this gives expected
results.

The reuse of the t1 alias is odd, but some databases require
an alias and apparently don't care if the same alias is used more
than once in separate parts or levels of the compound statement
structure.

This eliminates the special handling of compounds in the PostgreSQL
adapter and adapters that don't support INTERSECT or EXCEPT.  Now,
all adapters handle things the same way via the subselect approach.

I originally thought PostgreSQL would be faster if I used
parantheses instead of subselects, but it uses the same query
plan either way, so there was no point keeping the old code.

While here, add some documentation for some literal_* dataset
methods in the PostgreSQL adapter.