Support eager graphing for model tables that lack primary keys
Before, it was impossible to use a model that lacked a primary key
in an eager graph statement (either as the main table or one of
the associated tables.  However, it was possible to use eager
graphing for models that had primary keys but where the primary key
field was not selected.  Thus it is fairly obvious that a primary
key isn't required for it to work.

The reason it didn't work was that Model#pk raises an error if the
model doesn't have a primary key, and the eager graphing code calls
that on all models in order to build an identity map.

My original fix for this involved adding a new mapping of primary
keys for each table alias and passing that through the dependency
recurser.  That worked fine, but I didn't like the additional
complexity.

This implementation just adds Model#pk_or_nil that just returns
nil if the model doesn't have a primary key, instead of raising an
error.  I'm guessing this performs better, but it also is cleaner
and less invasive, which is the reason I chose it.

While here, use the ternary operator instead of a case statement
in Model#pk.